Move providers from js to ts.
