+++ src/providers/index.ts
@@ -3,11 +3,11 @@ import helpScout from './helpScout'
 import intercom from './intercom'
 import messenger from './messenger'
 
-const Providers = {
+const providers = {
   drift,
   helpScout,
   intercom,
   messenger
 }
 
-export default Providers
+export default providers
